fix: inline templated classes
